chore: admin 폴더 정리
