+++ src/main/java/core/global/admin/controller/AdminImageController.java
@@ -1,7 +1,6 @@
-package core.global.controller;
+package core.global.admin.controller;
 
 import core.global.entity.image.dto.SuspiciousImageResponse;
-import core.global.entity.image.entity.Image;
 import core.global.service.AdminImageService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
