+++ src/main/java/core/global/admin/controller/AdminMetricsViewController.java
@@ -1,4 +1,4 @@
-package core.global.controller;
+package core.global.admin.controller;
 
 import core.global.dto.AdminMetricsDto;
 import core.global.service.AdminMetricsService;
