+++ src/main/java/core/global/admin/controller/UserAdminViewController.java
@@ -1,4 +1,4 @@
-package core.global.controller;
+package core.global.admin.controller;
 
 import core.domain.chat.dto.RecentMessageDto;
 import core.domain.chat.entity.ChatRoom;
